// src/app/shared/models/child-progress.model.ts
export class ChildProgressModel {
    firstName: string;
    lastName: string;
    mothersLastName: string;
    birthDate: Date;
    numberOrder: string;
    folio: string;
    age: number;
    sex: string;
    date: Date;
    size: string;
    weight: string;
    pointA1: string;
    pointA2: string;
    pointA3: string;
    pointB1: string;
    pointB2: string;
    pointB3: string;
    pointC1: string;
    pointC2: string;
    pointC3: string;
    pointD1: string;
    pointD2: string;
    pointD3: string;
    totalA: number;
    totalB: number;
    totalC: number;
    totalD: number;
    total: number;


    constructor() {
        this.firstName = '';
        this.lastName = '';
        this.mothersLastName = '';
        this.birthDate = null;
        this.numberOrder = '';
        this.folio = '';
        this.age = null;
        this.sex = '';
        this.date = null;
        this.size =  '';
        this.weight = '';
        this.pointA1 = null;
        this.pointA2 = null;
        this.pointA3 = null;
        this.pointB1 = null;
        this.pointB2 = null;
        this.pointB3 = null;
        this.pointC1 = null;
        this.pointC2 = null;
        this.pointC3 = null;
        this.pointD1 = null;
        this.pointD2 = null;
        this.pointD3 = null;
        this.totalA = null;
        this.totalB = null;
        this.totalC = null;
        this.totalD = null;
        this.total = null;
    }
}